[server] small fix: время начала вызова (монитор процессов)

+++ server/src/main/java/lsfusion/server/remote/RemoteLoggerAspect.java
@@ -77,7 +77,7 @@ public class RemoteLoggerAspect {
         return lde != null && lde;
     }
 
-    public static Timestamp getDateTimeCall(int pid) {
+    public static Timestamp getDateTimeCall(long pid) {
         return dateTimeCallMap.get(pid);
     }
